added ready and route controller modules. script is gone. its over #495
