+++ src/js/ready.js
@@ -1,52 +1,17 @@
-(function (history) {
-  var pushState = history.pushState;
-  history.pushState = function (state) {
-    if (Funbox.active === "memory" && state !== "/") {
-      Funbox.resetMemoryTimer();
-    }
-    return pushState.apply(history, arguments);
-  };
-})(window.history);
-
-$(window).on("popstate", (e) => {
-  let state = e.originalEvent.state;
-  if (state == "" || state == "/") {
-    // show test
-    UI.changePage("test");
-  } else if (state == "about") {
-    // show about
-    UI.changePage("about");
-  } else if (state == "account" || state == "login") {
-    if (firebase.auth().currentUser) {
-      UI.changePage("account");
-    } else {
-      UI.changePage("login");
-    }
-  }
-});
+import * as ManualRestart from "./manual-restart-tracker";
+import Config, * as UpdateConfig from "./config";
+import * as Misc from "./misc";
+import * as VerificationController from "./verification-controller";
+import * as Settings from "./settings";
+import * as RouteController from "./route-controller";
+import * as UI from "./ui";
 
 ManualRestart.set();
 UpdateConfig.loadFromCookie();
 Misc.getReleasesFromGitHub();
 
-let mappedRoutes = {
-  "/": "pageTest",
-  "/login": "pageLogin",
-  "/settings": "pageSettings",
-  "/about": "pageAbout",
-  "/account": "pageAccount",
-  "/verify": "pageTest",
-};
-
-function handleInitialPageClasses(el) {
-  $(el).removeClass("hidden");
-  $(el).addClass("active");
-}
-
 $(document).ready(() => {
-  handleInitialPageClasses(
-    $(".page." + mappedRoutes[window.location.pathname])
-  );
+  RouteController.handleInitialPageClasses(window.location.pathname);
   if (window.location.pathname === "/") {
     $("#top .config").removeClass("hidden");
   }
